pyproject.toml
[project]
name = "spec-kitty-runtime"
version = "0.1.0a0"
description = "Canonical mission runtime for deterministic next-step planning"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Spec Kitty Contributors" }
]
dependencies = [
  "pydantic>=2.0,<3.0",
  "pyyaml>=6.0",
  "jsonschema>=4.0",
  "spec-kitty-events==2.2.0"
]

[project.optional-dependencies]
test = [
  "pytest>=8.0"
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
